split up genesis (#2289)
* split out the delta phase

* split out modules
